requirements.txt
python-dotenv==1.0.0
llama-index==0.9.10
html2text==2020.1.16
beautifulsoup4==4.12.2 
pinecone-client==2.2.4
nltk==3.8.1 
streamlit==1.29.0
